Update de.json (#195)
I have replaced the umlauts Ä, Ü and Ö with ae, ue and oe. There was a mistake on the character sheet. You could not enter values for strength, agility and toughness. The dice function for these values could not be carried out either as the code was interpreted incorrectly due to the umlauts. I have tested this locally. DR was exchanged for SG in German is “SG = Schweregrat” that fits better. HP was exchanged for LP in German “LP = Lebenspunkte”. “AUSSTRAHLUNG” was exchanged for 'Präsenz', which fits and describes the attribute better.  Missing translations have been made and the existing ones grammatically improved. I hope you release the file quickly so that the German MÖRG BORG community can also use the character sheet properly.